Implementing Auth

+++ app/Http/Controllers/ChirpController.php
@@ -2,11 +2,15 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 use App\Models\Chirp;
+use App\Models\User;
+use Illuminate\Support\Facades\Auth;
 
 class ChirpController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
@@ -37,6 +41,7 @@ class ChirpController extends Controller
         ]);
 
         Chirp::create([
+            "user_id" => Auth::id(),
             "message" => $validated["message"]
         ]);
 
@@ -56,6 +61,8 @@ class ChirpController extends Controller
      */
     public function edit(Chirp $chirp)
     {
+        $this->authorize("update", $chirp);
+
         return view("chirps.edit", compact("chirp"));
     }
 
@@ -64,6 +71,8 @@ class ChirpController extends Controller
      */
     public function update(Request $request, Chirp $chirp)
     {
+        $this->authorize("update", $chirp);
+
         $validated = $request->validate([
             "message" => "required|string|max:255|min:2",
         ], [
@@ -81,7 +90,7 @@ class ChirpController extends Controller
      */
     public function destroy(Chirp $chirp)
     {
-        // $this->authorize("delete", $chirp);
+        $this->authorize("delete", $chirp);
 
         $chirp->delete();
